pyproject.toml
[project]
name = "tp-cookie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.2",
    "pillow>=12.1.0",
    "scipy>=1.17.0",
]
